requirements.txt
torch>=2.0.0
transformers>=4.36.0
accelerate>=0.25.0
peft>=0.7.0
datasets>=2.14.0
pillow>=10.0.0
gradio>=4.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
huggingface_hub>=0.19.0
wandb>=0.16.0
sentencepiece>=0.1.99
protobuf>=3.20.0
